requirements.txt
line-bot-sdk==3.*
python-dotenv==0.21.1
Flask==2.2.2
opencc-python-reimplemented==0.1.4
beautifulsoup4==4.11.2
youtube-transcript-api==0.5.0
pymongo==4.3.3
gspread
google-auth
pytz